% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cats.R
\docType{data}
\name{cats}
\alias{cats}
\title{Cat names}
\format{A data frame with 29421 rows and 37 variables:
\describe{
  \item{name}{Cat name}
}}
\usage{
cats
}
\description{
This is a dataset full of adorable and sometimes bizarre cat names.
}
\keyword{datasets}
